% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocols.r
\name{add_true}
\alias{add_true}
\title{Add true data into data frame containing null data sets.}
\usage{
add_true(samples, true, pos)
}
\description{
Add true data into data frame containing null data sets.
}
\keyword{internal}
