\name{all_sample}
\alias{all_sample}
\title{Calculates the mean distance for all replicates}
\usage{
all_sample(dat, repl, no.samp, method)
}
\arguments{
  \item{dat}{observed data set using}

  \item{repl}{the number of replicates}

  \item{no.samp}{number of samples}

  \item{method}{null generating mechanism}
}
\description{
  Calculates the mean distance for all replicates
}
