\name{decrypt}
\alias{decrypt}
\title{Use decrypt to reveal the position of the real data.}
\usage{
  decrypt(...)
}
\arguments{
  \item{...}{character vector to decrypt}
}
\description{
  The real data position is encrypted by the lineup
  function, and writes this out as a text string. Decrypt,
  decrypts this text string to reveal which where the real
  data is.
}
\examples{
decrypt("0uXR2p rut L2O2")
}

