% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{electoral}
\alias{electoral}
\title{Polls and election results from the 2012 US Election}
\format{
A list with two data frames:
polls is a data frame of 51 rows and 4 variables
\describe{
\item{State}{State name}
\item{Electoral.vote}{Number of electoral votes in the 2012 election}
\item{Margin}{Margin between the parties with the highest number of votes and second highest number of votes.
These margins are based on polls.}
\item{Democrat}{logical vector True, if the democratic party is  the majority party in this state. }
}
`election` is a data frame of 51 rows and 5 variables
\describe{
\item{State}{State name}
\item{Candidate}{character string of the winner: Romney or Obama}
\item{Electoral.vote}{Number of electoral votes in the 2012 election}
\item{Margin}{Margin between the parties with the highest number of votes and second highest number of votes.
These margins are based on the actual election outcome}
\item{Democrat}{logical vector True, if the democratic party is  the majority party in this state. }
}
}
\description{
Polls and election results from the 2012 US Election
}
