\name{mean.samp}
\alias{mean.samp}
\title{Calculates the mean distance for each plot in the lineup.}
\usage{
mean.samp(dat, no.samp, method)
}
\arguments{
  \item{dat}{lineup data}

  \item{no.samp}{number of samples}

  \item{method}{null generating mechanism}
}
\description{
  Calculates the mean distance for each plot in the lineup.
}
