\name{mean_met}
\alias{mean_met}
\title{Calculates the mean distance between the null datasets}
\usage{
mean_met(dat, no.samp, method)
}
\arguments{
  \item{dat}{observed data set using}

  \item{no.samp}{number of samples}

  \item{method}{null generating mechanism}
}
\description{
  Calculates the mean distance between the null datasets
}
