% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmet.r
\name{null_gen}
\alias{null_gen}
\title{Computing th distance for the null plots}
\usage{
null_gen(lineup.dat, null, met, method, m, dist.arg)
}
\description{
Computing th distance for the null plots
}
\keyword{internal}
