% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-model.r
\name{resid_boot}
\alias{resid_boot}
\title{Bootstrap residuals.}
\usage{
resid_boot(model, data)
}
\arguments{
\item{model}{to extract residuals from}

\item{data}{used to fit model}
}
\description{
For use with \code{\link{null_lm}}
}
