% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resid_pboot}
\alias{resid_pboot}
\title{Parametric bootstrap residuals.}
\usage{
resid_pboot(model, data)
}
\arguments{
\item{model}{to extract residuals from}

\item{data}{used to fit model}
}
\description{
For use with \code{\link{null_lm}}
}

