% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resid_sigma}
\alias{resid_sigma}
\title{Residuals simulated by a normal model, with specified sigma}
\usage{
resid_sigma(model, data, sigma = 1)
}
\arguments{
\item{model}{to extract residuals from}

\item{data}{used to fit model}

\item{sigma,}{a specific sigma to model}
}
\description{
For use with \code{\link{null_lm}}
}

