% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_strip}
\alias{theme_strip}
\title{A theme to minimally strip away the context}
\usage{
theme_strip()
}
\description{
Note this is not a complete theme hence why there are no arguments.
}
\examples{
library(ggplot2)
ggplot(cars, aes(dist, speed)) + theme_strip()

}
