% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.r
\name{visual_power}
\alias{visual_power}
\title{Power calculations.}
\usage{
visual_power(data, m = 20)
}
\arguments{
\item{data}{summary of the results, containing columns id, pic_id, response, detected}

\item{m}{size of the lineup}
}
\value{
vector of powers for each pic_id
}
\description{
This function simply counts the proportion of people who selected the data plot,
in a set of lineups. It adjusts for multiple picks by the same individual, by weighting
by the total number of choices.
}
\examples{
data(turk_results)
visual_power(turk_results)
}
