% Generated by roxygen2 (4.0.2): do not edit by hand
\name{box_dist}
\alias{box_dist}
\title{Distance based on side by side Boxplots for two levels}
\usage{
box_dist(X, PX)
}
\arguments{
\item{X}{a data.frame with one factor variable and one continuous
variable}

\item{PX}{a data.frame with one factor variable and one continuous
variable}
}
\value{
distance between X and PX
}
\description{
Assuming there are only two groups, the first quartile, median and third quartile
is calculated for each group of data X. The absolute difference between these
statistics between the two groups are then calculated. Same is done for data PX.
Finally an euclidean distance is calculated between the absolute differences of
X and PX.
}
\examples{
if(require('dplyr')) {with(mtcars, box_dist(data.frame(as.factor(am), mpg),
data.frame(as.factor(sample(am)), mpg)))}
}

