% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distplot}
\alias{distplot}
\title{Plotting the distribution of the distance measure}
\usage{
distplot(dat, m = 20)
}
\arguments{
\item{dat}{output from \code{\link{distmet}}}

\item{m}{the number of plots in the lineup; m = 20 by default}
}
\description{
The distribution of the distance measure is plotted with the distances for
the null plots and true plot overlaid.
}
\examples{
if(require('dplyr')){
distplot(distmet(lineup(null_permute('mpg'), mtcars, pos = 1), var = c('mpg',
'wt'), 'reg_dist', null_permute('mpg'), pos = 1, repl = 100, m = 8), m = 8) }
}

