\name{reg_dist}
\alias{reg_dist}
\title{Distance based on the regression parameters}
\usage{
  reg_dist(X, PX, X.bin = 1, Y.bin = X.bin)
}
\arguments{
  \item{X}{a data.frame with two variables, the first
  column giving the explanatory variable and the second
  column giving the response variable}

  \item{PX}{another data.frame with two variables, the
  first column giving the explanatory variable and the
  second column giving the response variable}
}
\value{
  distance between X and PX
}
\description{
  Distance based on the regression parameters
}

