\name{resid_pboot}
\alias{resid_pboot}
\title{Parametric bootstrap residuals.}
\usage{resid_pboot(model, data)}

\description{
  Parametric bootstrap residuals.
}

\details{
  For use with \code{\link{null_lm}}
}
\arguments{
  \item{model}{to extract residuals from}
  \item{data}{used to fit model}
}
