% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-model.r
\name{resid_sigma}
\alias{resid_sigma}
\title{Residuals simulated by a normal model, with specified sigma}
\usage{
resid_sigma(model, data, sigma = 1)
}
\arguments{
\item{model}{to extract residuals from}

\item{data}{used to fit model}

\item{sigma, }{a specific sigma to model}
}
\description{
For use with \code{\link{null_lm}}
}
