% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.r
\name{uni_dist}
\alias{uni_dist}
\title{Distance for univariate data}
\usage{
uni_dist(X, PX)
}
\arguments{
\item{X}{a data.frame where the first column is only used}

\item{PX}{another data.frame where the first column is only used}
}
\value{
distance between X and PX
}
\description{
The first four moments is calculated for data X and data PX. An euclidean distance
is calculated between these moments for X and PX.
}
\examples{
if(require('moments')){uni_dist(rnorm(100), rpois(100, 2))}
}
